deploy/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.9.0
